Remove unnecessary "tid" fields from odb and data shared memory, sort out confusion between pid and tid.
SVN revision: 4794

+++ src/system.c
@@ -937,7 +937,7 @@ INT ss_gettid(void)
 
 #elif defined OS_DARWIN
 
-   return pthread_self();
+   return (int)pthread_self();
 
 #elif defined OS_CYGWIN
